requirements.txt
prometheus-client
pydantic
pyyaml
